Fixed the installation instructions
The README talked a bit too much about Paw, which is not what this library is going to be used for by the community at large.

It also focused on compile instructions over recommending npm/yarn installation. ALSO the compile instructions were invalid, and talked about make install (doesnt exist!).